pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.black]
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
(
  ^/external/
  | ^/foo.py    # exclude a file named foo.py in the root of the project
  | .*_pb2.py  # exclude autogenerated Protocol Buffer files anywhere in the project
)
'''
include = '\.pyi?$'
line-length = 79

[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "async-python-sprint-4"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
alembic = "^1.12.0"
asyncpg = "^0.28.0"
databases = {extras = ["postgresql"], version = "^0.8.0"}
ecs-logging = "^2.1.0"
fastapi = "^0.103.1"
httpx = "^0.25.0"
hypercorn = "^0.14.4"
orjson = "3.8.0"
psycopg2-binary = "^2.9.7"
pydantic = "^2.3.0"
pydantic-settings = "^2.0.3"
pyodbc = "^4.0.39"
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"
python = "^3.10"
sqlalchemy-utils = "^0.41.1"
toml-sort = "^0.23.1"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = [".", "src"]
